[ch817] adds spec for patient milestone methods

+++ db/migrate/20160825145834_add_category_and_age_of_patient_in_months_to_link_preview.rb
@@ -1,6 +1,6 @@
 class AddCategoryAndAgeOfPatientInMonthsToLinkPreview < ActiveRecord::Migration
   def change
     add_column :link_previews, :category, :string
-    add_column :link_previews, :age_of_patient_in_months, :float
+    add_column :link_previews, :age_of_patient_in_months, :integer
   end
 end
